// repos/sbs-draft-web-main/utils/types/types.ts
export type OwnerProps = {
    availableCredit: number
    availableEthCredit: number
    blueCheckEmail: string
    hasW9: boolean
    isBlueCheckVerified: boolean
    leagues: LeagueProps[]
    numWithdrawals: number
    pendingCredit: number
    withdrawnAmount: {
        2022: number
    }
    pfp: {
        imageUrl: string
        nftContract: string
        displayName: string
    }
}

export type LeagueProps = {
    leagueId: string
    cardId: string
}

export type NFTProps = {
    identifier: string
    collection: string
    contract: string
    token_standard: string
    name: string
    description: string
    image_url: string
    is_disabled: boolean
    is_nsfw: boolean
    metadata_url: string
}